Front-end Add User for Admin

+++ front-end/src/pages/AdminProducts/AdminProducts.jsx
@@ -14,7 +14,7 @@ import {
 import Pagination from "../../components/Pagination/Pagination";
 import Loader from "../../components/Loader/Loader";
 import Toastify from "../../components/Toastify/Toastify";
-
+import Swal from "sweetalert2";
 import { Link, useOutletContext } from "react-router-dom";
 
 import "./admin-products.css";
@@ -94,33 +94,45 @@ const AdminProducts = () => {
   };
 
   const deleteProduct = (id) => {
-    fetch(`http://localhost:8000/api/v1/product/${id}`, {
-      credentials: "include",
-      method: "DELETE",
-      headers: {
-        Authorization: `Bearer ${localStorage.getItem("authToken")}`,
-      },
-    })
-      .then((res) => res.json())
-      .then((json) => {
-        if (json.status === "success") {
-          Toastify(
-            "success",
-            "bottom-right",
-            "Successfully removed that product from the store"
-          );
-          if (productList.length === 1 && currentPage > 1) {
-            setCurrentPage(currentPage - 1);
-          }
-          setDel(!del);
-        } else {
-          Toastify(
-            "error",
-            "bottom-right",
-            "Failed to remove that product from the store!!! Please try again"
-          );
-        }
-      });
+    Swal.fire({
+      title: "Are you sure you want to do this?",
+      text: "You will never see this product in your store again!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, remove it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch("http://localhost:8000/api/v1/product/" + id, {
+          credentials: "include",
+          method: "DELETE",
+          headers: {
+            Authorization: `Bearer ${localStorage.getItem("authToken")}`,
+          },
+        })
+          .then((res) => res.json())
+          .then((json) => {
+            if (json.status === "success") {
+              Swal.fire({
+                title: "Product deleted!",
+                text: "Successfully removed that product from your store",
+                icon: "success",
+              });
+              if (productList.length === 1 && currentPage > 1) {
+                setCurrentPage(currentPage - 1);
+              }
+              setDel(!del);
+            } else {
+              Swal.fire({
+                title: "Error while removing!",
+                text: "There's some error!!! Please try again later",
+                icon: "error",
+              });
+            }
+          });
+      }
+    });
   };
   return (
     <>
